Finally resolve the libcxx contagion issue, fixing builds outside macOS.
This requires using the version 2 resolver, which is something you need to set on your crate that *uses* psilo-text. As such, we include a dummy dependency, which, on the version 1 resolver, raises a helpful compiler error message, and on the version 2 resolver, works fine. So... do what the message says and you'll be fine.